Fix some bugs and add some stuff that I forgot earlier

+++ editor/src/main/java/nl/knokko/customitems/editor/menu/edit/collection/DedicatedCollectionEdit.java
@@ -1,15 +1,18 @@
 package nl.knokko.customitems.editor.menu.edit.collection;
 
 import nl.knokko.customitems.model.ModelValues;
+import nl.knokko.customitems.util.Checks;
 import nl.knokko.gui.color.GuiColor;
 import nl.knokko.gui.component.GuiComponent;
 import nl.knokko.gui.component.image.SimpleImageComponent;
 import nl.knokko.gui.component.menu.GuiMenu;
+import nl.knokko.gui.component.text.EagerTextEditField;
 import nl.knokko.gui.component.text.dynamic.DynamicTextButton;
 import nl.knokko.gui.component.text.dynamic.DynamicTextComponent;
 import nl.knokko.gui.texture.loader.GuiTextureLoader;
 
 import java.awt.image.BufferedImage;
+import java.util.Locale;
 import java.util.function.Function;
 import java.util.function.Supplier;
 
@@ -23,16 +26,19 @@ public abstract class DedicatedCollectionEdit<V extends ModelValues, R extends S
     protected final Function<V, String> attemptAddModel;
 
     protected final ItemList itemList;
+    protected final EagerTextEditField searchField;
     protected final DynamicTextComponent errorComponent;
 
     public DedicatedCollectionEdit(
             GuiComponent returnMenu, Iterable<R> liveCollection, Function<V, String> attemptAddModel
     ) {
+        Checks.notNull(returnMenu);
         this.returnMenu = returnMenu;
         this.liveCollection = liveCollection;
         this.attemptAddModel = attemptAddModel;
 
         this.itemList = new ItemList();
+        this.searchField = new EagerTextEditField("", EDIT_BASE, EDIT_ACTIVE, newText -> itemList.refresh());
         this.errorComponent = new DynamicTextComponent("", ERROR);
     }
 
@@ -41,7 +47,14 @@ public abstract class DedicatedCollectionEdit<V extends ModelValues, R extends S
         addComponent(new DynamicTextButton("Back", CANCEL_BASE, CANCEL_HOVER, () ->
                 state.getWindow().setMainComponent(returnMenu)
         ), 0.025f, 0.7f, 0.175f, 0.8f);
-        addComponent(itemList, 0.25f, 0f, 1f, 0.9f);
+
+        addComponent(
+                new DynamicTextComponent("Search:", LABEL),
+                0.025f, 0.6f, 0.15f, 0.7f
+        );
+        addComponent(searchField, 0.025f, 0.5f, 0.28f, 0.6f);
+
+        addComponent(itemList, 0.3f, 0f, 1f, 0.9f);
         addComponent(errorComponent, 0.05f, 0.9f, 0.95f, 1f);
     }
 
@@ -91,52 +104,53 @@ public abstract class DedicatedCollectionEdit<V extends ModelValues, R extends S
 
             for (R modelReference : liveCollection) {
                 V model = modelReference.get();
-
-                float maxY = minY + 0.1f;
-                BufferedImage icon = getModelIcon(model);
-                if (icon != null) {
-                    addComponent(new SimpleImageComponent(textureLoader.loadTexture(icon)), 0f, minY, 0.125f, maxY);
-                }
-
                 String label = getModelLabel(model);
+                if (label.toLowerCase(Locale.ROOT).contains(searchField.getText().toLowerCase(Locale.ROOT))) {
 
-                addComponent(
-                        new DynamicTextComponent(label, LABEL),
-                        minTextX, minY, Math.min(0.6f, minTextX + 0.05f * label.length()), maxY
-                );
+                    float maxY = minY + 0.1f;
+                    BufferedImage icon = getModelIcon(model);
+                    if (icon != null) {
+                        addComponent(new SimpleImageComponent(textureLoader.loadTexture(icon)), 0f, minY, 0.125f, maxY);
+                    }
 
-                if (canEditModel(model)) {
-                    addComponent(new DynamicTextButton("Edit", BUTTON, HOVER, () ->
-                            state.getWindow().setMainComponent(createEditMenu(modelReference))
-                    ), 0.61f, minY, 0.72f, maxY);
-                }
+                    addComponent(
+                            new DynamicTextComponent(label, LABEL),
+                            minTextX, minY, Math.min(0.6f, minTextX + 0.05f * label.length()), maxY
+                    );
 
-                CopyMode copyMode = getCopyMode(modelReference);
-                if (copyMode != CopyMode.DISABLED) {
-                    addComponent(new DynamicTextButton("Copy", BUTTON, HOVER, () -> {
-                        if (copyMode == CopyMode.SEPARATE_MENU) {
-                            state.getWindow().setMainComponent(createCopyMenu(modelReference));
-                        } else if (copyMode == CopyMode.INSTANT) {
-                            String error = attemptAddModel.apply(model);
+                    if (canEditModel(model)) {
+                        addComponent(new DynamicTextButton("Edit", BUTTON, HOVER, () ->
+                                state.getWindow().setMainComponent(createEditMenu(modelReference))
+                        ), 0.61f, minY, 0.72f, maxY);
+                    }
+
+                    CopyMode copyMode = getCopyMode(modelReference);
+                    if (copyMode != CopyMode.DISABLED) {
+                        addComponent(new DynamicTextButton("Copy", BUTTON, HOVER, () -> {
+                            if (copyMode == CopyMode.SEPARATE_MENU) {
+                                state.getWindow().setMainComponent(createCopyMenu(modelReference));
+                            } else if (copyMode == CopyMode.INSTANT) {
+                                String error = attemptAddModel.apply(model);
+                                if (error == null) {
+                                    refresh();
+                                } else {
+                                    errorComponent.setText(error);
+                                }
+                            }
+                        }), 0.73f, minY, 0.83f, maxY);
+                    }
+                    if (canDeleteModels()) {
+                        addComponent(new DynamicTextButton("Delete", QUIT_BASE, QUIT_HOVER, () -> {
+                            String error = deleteModel(modelReference);
                             if (error == null) {
                                 refresh();
                             } else {
                                 errorComponent.setText(error);
                             }
-                        }
-                    }), 0.73f, minY, 0.83f, maxY);
+                        }), 0.84f, minY, 0.99f, maxY);
+                    }
+                    minY -= 0.1f;
                 }
-                if (canDeleteModels()) {
-                    addComponent(new DynamicTextButton("Delete", QUIT_BASE, QUIT_HOVER, () -> {
-                        String error = deleteModel(modelReference);
-                        if (error == null) {
-                            refresh();
-                        } else {
-                            errorComponent.setText(error);
-                        }
-                    }), 0.84f, minY, 0.99f, maxY);
-                }
-                minY -= 0.1f;
             }
         }
